fix(lyrics): rewrite Phase 2.65 doc + extract tests sibling
CI lint failed on doc-lazy-continuation in absorb_leading_unmatched
docstring (em-dashes were misread as list-item continuations). Rewrite
docstring as plain prose without dashes/bullets.

Move four Phase 2.65 unit tests into a new sibling
text_reference_merge_absorb_tests.rs so the parent tests file stays
under the 1000-line cap.
